requirements.txt
Django>=5.0,<6.0
mysqlclient; platform_system != 'Windows'
PyMySQL; platform_system == 'Windows'
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=22.0
whitenoise>=6.6
